feat: moves `WalletConnect2Session` in separate file
